fixe negatif

+++ srcs/malloc.c
@@ -6,7 +6,7 @@
 /*   By: qduperon <qduperon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/27 13:37:24 by qduperon          #+#    #+#             */
-/*   Updated: 2019/02/27 14:17:54 by qduperon         ###   ########.fr       */
+/*   Updated: 2019/02/27 15:54:34 by qduperon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_malloc(size_t size)
 {
 	void *addr;
 
-	if (size < 1)
+	if ((int)size < 1)
 		return (NULL);
 	if ((addr = get_from_gmap(size)) != NULL)
 		return (addr);
